fixing previous commit

+++ src/ags/audio/task/recall/ags_apply_sequencer_length.h
@@ -40,7 +40,7 @@ struct _AgsApplySequencerLength
 
   GObject *gobject;
 
-  guint length;
+  gdouble length;
 };
 
 struct _AgsApplySequencerLengthClass
@@ -51,6 +51,6 @@ struct _AgsApplySequencerLengthClass
 GType ags_apply_sequencer_length_get_type();
 
 AgsApplySequencerLength* ags_apply_sequencer_length_new(GObject *gobject,
-							guint length);
+							gdouble length);
 
 #endif /*__AGS_APPLY_SEQUENCER_LENGTH_H__*/
